requirements.txt
crewai
# crewai-tools
icalendar
requests
twilio
python-dotenv